Forcing header styles to automatically apply.
It's better to have header style classes that you can apply as
you wish, rather than have the styles tied directly to the header
tags. However, on sites that use markup we put the onus on the
markup writer to add classes each time they want to use a header.
This is undesirable, and so we apply styles to the markup
directly...