Refactoring buttons to darken on hover.
This also introduces a change in the markup paradigm. We should
add the button class to get the button structure, then a button
color class to get the appropriate colors. This reduces the
number of css rules our sass generates and consequently lowers
the size of our final css file.